Aurum_Infinity_AI/requirements.txt
flask==3.1.3
flask-limiter==4.1.1
markdown==3.10.2
requests==2.32.5
google-genai==1.65.0
python-dotenv==1.2.2
gunicorn==25.1.0
PyYAML==6.0.3
